[#34] Fix stuck delay jobs from index-purge edge cases
Under certain conditions, asynchronous jobs were accumulating in
the iRODS delay queue. Here, we address that by logging a 404 response from
elasticsearch during an AVU purge and returning a success code,
and analogously trapping and nullifying the exception caused
when an indexed data object is removed and the delayed purge task
is unable to locate it.

For the present, in the case of the data object, the indexed
contents still remain; see [#37].