[#44] skip tests using a Python3 virtual env
Tests that rely on an python-irodsclient(PRC)-based client presently
try to arrange a virtual environment for this purpose.

Until changes are made in the test hook which use purely the OS
package manager to install python3-virtualenv (in C7, possible only
with the EPEL repository enabled), we will skip such tests in the
CI test runs.

These tests can, however, be run at the command-line by setting the
environment variable MANUALLY_TEST_INDEXING_PLUGIN to a nonzero-length
string not having a value case-insensitively equivalent to "n", "no",
"f", "false", or "0".